Wrap the game in React.memo to prevent renderering
